test: extract provider sections to tests/providers/ (batch E, #1440)
