+++ test-all.mjs
@@ -5568,168 +5568,6 @@ if (!fileExists('interview-prep/sessions/README.md')) {
   }
 }
 
-console.log('\n53. Provider — successfactors (SAP RMK tile parser)');
-
-try {
-  const sf = (await import(pathToFileURL(join(ROOT, 'providers/successfactors.mjs')).href)).default;
-  const { parseTiles, cityFromSlug } = await import(pathToFileURL(join(ROOT, 'providers/successfactors.mjs')).href);
-
-  if (sf.id === 'successfactors') pass('successfactors.id is "successfactors"');
-  else fail(`successfactors.id is ${JSON.stringify(sf.id)}`);
-
-  // detect() — literal SF hosts auto-claim; branded RMK hosts (jobs.zf.com) do
-  // NOT (they carry no "successfactors" string and rely on explicit provider:).
-  if (sf.detect({ name: 'X', careers_url: 'https://acme.successfactors.eu/careers' })) {
-    pass('successfactors.detect() claims a *.successfactors.eu URL');
-  } else {
-    fail('successfactors.detect() should claim *.successfactors.eu');
-  }
-  if (sf.detect({ name: 'X', api: 'https://company.jobs2web.com/x' })) {
-    pass('successfactors.detect() claims a jobs2web.com URL');
-  } else {
-    fail('successfactors.detect() should claim jobs2web.com');
-  }
-  if (sf.detect({ name: 'ZF', careers_url: 'https://jobs.zf.com' }) === null) {
-    pass('successfactors.detect() returns null for a branded RMK host (needs explicit provider:)');
-  } else {
-    fail('successfactors.detect() must not auto-claim branded hosts');
-  }
-
-  // cityFromSlug — recover the city prefix from an RMK /job/{City}-{Title}-{code}/ slug.
-  if (cityFromSlug('/job/Hyderabad-Specialist-Low-Level-Driver-Development-TG-500032/1399717233/', 'Specialist -Low Level Driver Development') === 'Hyderabad') {
-    pass('cityFromSlug extracts a single-word city');
-  } else {
-    fail(`cityFromSlug single-word wrong: ${cityFromSlug('/job/Hyderabad-Specialist-Low-Level-Driver-Development-TG-500032/1399717233/', 'Specialist -Low Level Driver Development')}`);
-  }
-  // Multi-word city (Levallois-Perret) — anchoring on the title's first two
-  // words means the full city prefix survives, not just the first token.
-  if (cityFromSlug('/job/Levallois-Perret-Data-Management-Engagement-Architect-92300/1400945133/', 'Data Management Engagement Architect') === 'Levallois Perret') {
-    pass('cityFromSlug extracts a multi-word city');
-  } else {
-    fail(`cityFromSlug multi-word wrong: ${cityFromSlug('/job/Levallois-Perret-Data-Management-Engagement-Architect-92300/1400945133/', 'Data Management Engagement Architect')}`);
-  }
-  // Accented title (Ingénieur) — unicode word matching keeps the anchor intact.
-  if (cityFromSlug('/job/Massy-Ing%C3%A9nieur-Commercial-91743/1351400755/', 'Ingénieur Commercial') === 'Massy') {
-    pass('cityFromSlug handles accented (unicode) titles');
-  } else {
-    fail(`cityFromSlug accented wrong: ${cityFromSlug('/job/Massy-Ing%C3%A9nieur-Commercial-91743/1351400755/', 'Ingénieur Commercial')}`);
-  }
-
-  // parseTiles — a compact fragment covering the three things that bit during
-  // development: the city-value div (not its "City" label), an &amp; in the
-  // data-url path, a slug fallback when no city div is rendered, an entity in
-  // the title, desktop/mobile duplication collapsed to one <li>, and a
-  // title-less tile that must be dropped.
-  const jobBase = 'https://jobs.example.com';
-  const fragment = `
-    <ul>
-      <li class="job-tile job-id-111 job-row-index-1" data-url="/job/Schweinfurt-Ferienarbeiter-97421/111/">
-        <a class="jobTitle-link fontcolorx" href="/job/Schweinfurt-Ferienarbeiter-97421/111/">Ferienarbeiter (m&#47;w&#47;d)</a>
-        <div id="job-111-desktop-section-city" class="section-field city">
-          <span id="job-111-desktop-section-city-label" aria-describedby="job-111-desktop-section-city-value" class="section-label sr-only">City</span>
-          <div id="job-111-desktop-section-city-value">Schweinfurt                 </div>
-        </div>
-      </li>
-      <li class="job-tile job-id-222 job-row-index-2" data-url="/job/Palo-Alto-Program-&amp;-Release-Manager-CA-94304/222/">
-        <a class="jobTitle-link fontcolorx" href="/x">Program &amp; Release Manager</a>
-      </li>
-      <li class="job-tile job-id-333 job-row-index-3" data-url="/job/no-title/333/">
-      </li>
-    </ul>`;
-  const parsed = parseTiles(fragment, jobBase);
-
-  if (parsed.length === 2) pass('parseTiles returns 2 jobs (title-less tile dropped)');
-  else fail(`parseTiles returned ${parsed.length} jobs, expected 2`);
-
-  const j1 = parsed.find((j) => j.url.includes('/111/'));
-  if (j1 && j1.title === 'Ferienarbeiter (m/w/d)') pass('parseTiles decodes entity in title');
-  else fail(`parseTiles title wrong: ${JSON.stringify(j1 && j1.title)}`);
-  if (j1 && j1.location === 'Schweinfurt') pass('parseTiles reads the city-value div, not the "City" label');
-  else fail(`parseTiles city wrong: ${JSON.stringify(j1 && j1.location)}`);
-  if (j1 && j1.url === 'https://jobs.example.com/job/Schweinfurt-Ferienarbeiter-97421/111/') pass('parseTiles builds an absolute URL from data-url');
-  else fail(`parseTiles url wrong: ${JSON.stringify(j1 && j1.url)}`);
-
-  const j2 = parsed.find((j) => j.url.includes('/222/'));
-  if (j2 && j2.url === 'https://jobs.example.com/job/Palo-Alto-Program-&-Release-Manager-CA-94304/222/') {
-    pass('parseTiles decodes &amp; in the data-url path');
-  } else {
-    fail(`parseTiles &amp; url wrong: ${JSON.stringify(j2 && j2.url)}`);
-  }
-  if (j2 && j2.location === 'Palo Alto') pass('parseTiles falls back to slug city when no city div is present');
-  else fail(`parseTiles slug-fallback city wrong: ${JSON.stringify(j2 && j2.location)}`);
-
-  // Empty fragment (MTU's zero-req case) → no jobs, no throw.
-  if (parseTiles('<!DOCTYPE html>', jobBase).length === 0) pass('parseTiles returns [] for an empty fragment');
-  else fail('parseTiles should return [] for an empty fragment');
-
-  // ── CSB (Career Site Builder) strategy — JSON jobs API ────────────────────
-  const { extractLocales, parseCsbDate, cleanCsbLocation, parseCsbJobs } =
-    await import(pathToFileURL(join(ROOT, 'providers/successfactors.mjs')).href);
-
-  // extractLocales — pull the language-switcher locales from a /search/ page,
-  // deduped and priority-ordered (de_DE, en_US first; then alphabetical).
-  const switcherHtml =
-    '<a href="/search/?q=&amp;startrow=0&amp;locale=fr_FR">FR</a>' +
-    '<a href="/search/?q=&amp;startrow=0&amp;locale=en_US">EN</a>' +
-    '<a href="/search/?q=&amp;startrow=0&amp;locale=de_DE">DE</a>' +
-    '<a href="/search/?q=&amp;startrow=0&amp;locale=de_DE">DE dup</a>';
-  const locs = extractLocales(switcherHtml);
-  if (JSON.stringify(locs) === JSON.stringify(['de_DE', 'en_US', 'fr_FR'])) {
-    pass('extractLocales dedups and priority-orders (de_DE, en_US, then alpha)');
-  } else {
-    fail(`extractLocales wrong: ${JSON.stringify(locs)}`);
-  }
-  if (extractLocales('<p>no locales here</p>').length === 0) pass('extractLocales returns [] when the page carries none');
-  else fail('extractLocales should return [] for a page with no locale links');
-
-  // parseCsbDate — locale-dependent short date; separator infers field order.
-  if (parseCsbDate('6/18/26') === Date.UTC(2026, 5, 18)) pass('parseCsbDate reads US M/D/YY');
-  else fail(`parseCsbDate US wrong: ${parseCsbDate('6/18/26')}`);
-  if (parseCsbDate('20.11.23') === Date.UTC(2023, 10, 20)) pass('parseCsbDate reads European D.M.YY (dots)');
-  else fail(`parseCsbDate DE wrong: ${parseCsbDate('20.11.23')}`);
-  if (parseCsbDate('garbage') === undefined && parseCsbDate('13/40/99') === undefined && parseCsbDate('') === undefined) {
-    pass('parseCsbDate returns undefined for junk / out-of-range / empty');
-  } else {
-    fail('parseCsbDate should reject junk, out-of-range, and empty input');
-  }
-
-  // cleanCsbLocation — array of "City, CC, ZIP<br/>" strings → joined, stripped.
-  if (cleanCsbLocation(['Karlovy Vary, CZE, 36004<br/>']) === 'Karlovy Vary, CZE, 36004') pass('cleanCsbLocation strips trailing <br/>');
-  else fail(`cleanCsbLocation single wrong: ${JSON.stringify(cleanCsbLocation(['Karlovy Vary, CZE, 36004<br/>']))}`);
-  if (cleanCsbLocation(['Munich<br/>', 'Berlin<br/>']) === 'Munich / Berlin') pass('cleanCsbLocation joins multiple locations with " / "');
-  else fail(`cleanCsbLocation multi wrong: ${JSON.stringify(cleanCsbLocation(['Munich<br/>', 'Berlin<br/>']))}`);
-  if (cleanCsbLocation(undefined) === '' && cleanCsbLocation([]) === '') pass('cleanCsbLocation tolerates missing/empty location');
-  else fail('cleanCsbLocation should return "" for missing/empty input');
-
-  // parseCsbJobs — map the {response:{…}} records; build {id}-{locale} URLs and
-  // sanitize the cosmetic slug (HTML entities, URL-structural chars).
-  const csbJson = {
-    totalJobs: 3,
-    jobSearchResult: [
-      { response: { id: '31099', unifiedStandardTitle: 'Analytical Lab Technician', unifiedUrlTitle: 'Analytical-Lab-Technician', jobLocationShort: ['Anyang, KOR, 14058<br/>'], unifiedStandardStart: '6/18/26' } },
-      { response: { id: '1283', unifiedStandardTitle: 'Senior Expert Mergers & Acquisitions (m/f/d)', unifiedUrlTitle: 'Senior-Expert-Mergers-&amp;-Acquisitions-%28mfd%29', jobLocationShort: ['Munich<br/>'], unifiedStandardStart: '4/21/26' } },
-      { response: { id: '', unifiedStandardTitle: 'No ID — dropped', unifiedUrlTitle: 'x' } },
-      { response: { id: '999', unifiedStandardTitle: '', unifiedUrlTitle: 'no-title-dropped' } },
-    ],
-  };
-  const csbCfg = { origin: 'https://jobs.example.com' };
-  const csbJobs = parseCsbJobs(csbJson, csbCfg, 'en_US');
-  if (csbJobs.length === 2) pass('parseCsbJobs drops records missing id or title');
-  else fail(`parseCsbJobs returned ${csbJobs.length}, expected 2`);
-  const c1 = csbJobs[0];
-  if (c1 && c1.url === 'https://jobs.example.com/job/Analytical-Lab-Technician/31099-en_US') pass('parseCsbJobs builds {origin}/job/{slug}/{id}-{locale}');
-  else fail(`parseCsbJobs url wrong: ${JSON.stringify(c1 && c1.url)}`);
-  if (c1 && c1.location === 'Anyang, KOR, 14058') pass('parseCsbJobs cleans jobLocationShort');
-  else fail(`parseCsbJobs location wrong: ${JSON.stringify(c1 && c1.location)}`);
-  if (c1 && c1.postedAt === Date.UTC(2026, 5, 18)) pass('parseCsbJobs sets postedAt from unifiedStandardStart');
-  else fail(`parseCsbJobs postedAt wrong: ${JSON.stringify(c1 && c1.postedAt)}`);
-  const c2 = csbJobs[1];
-  if (c2 && !/[?#&]|&amp;/.test(new URL(c2.url).pathname)) pass('parseCsbJobs sanitizes &amp; / URL-structural chars out of the slug');
-  else fail(`parseCsbJobs slug not sanitized: ${JSON.stringify(c2 && c2.url)}`);
-} catch (err) {
-  fail(`successfactors provider test threw: ${err.message}`);
-}
-
 // ── match-star.mjs — fixture story-bank + top match assertion ───────────────
 
 console.log('\n🧪 Testing match-star.mjs keyword scorer...');
@@ -5875,639 +5713,6 @@ try {
   fail(`prepare-application contract check crashed: ${e.message}`);
 }
 
-// ── 53. PROVIDER — WORKDAY ────────────────────────────────────────
-
-console.log('\n53. Provider — workday');
-
-try {
-  const workday = (await import(pathToFileURL(join(ROOT, 'providers/workday.mjs')).href)).default;
-  const { parseWorkdayResponse } = await import(pathToFileURL(join(ROOT, 'providers/workday.mjs')).href);
-
-  // Shared mock ctx shape for workday.fetch() calls below — only fetchJson varies per test.
-  // sleep is a no-op so retry-backoff delays don't slow the test suite down.
-  const mkWorkdayCtx = (fetchJson, extra = {}) => ({
-    transport: 'http',
-    fetchText: async () => { throw new Error('fetchText should not be called'); },
-    fetchJson,
-    sleep: async () => {},
-    ...extra,
-  });
-
-  if (workday.id === 'workday') pass('workday.id is "workday"');
-  else fail(`workday.id is ${JSON.stringify(workday.id)}`);
-
-  // detect() — valid Workday URLs
-  const hitUs = workday.detect({ name: 'Acme', careers_url: 'https://acme.wd12.myworkdayjobs.com/en-US/acme-jobs' });
-  if (hitUs && hitUs.url === 'https://acme.wd12.myworkdayjobs.com/wday/cxs/acme/acme-jobs/jobs') {
-    pass('workday.detect() resolves wd12 URL to CXS API endpoint');
-  } else {
-    fail(`workday.detect(wd12) returned ${JSON.stringify(hitUs)}`);
-  }
-
-  const hitNoLocale = workday.detect({ name: 'Test', careers_url: 'https://test.wd5.myworkdayjobs.com/TestBoard' });
-  if (hitNoLocale && hitNoLocale.url === 'https://test.wd5.myworkdayjobs.com/wday/cxs/test/TestBoard/jobs') {
-    pass('workday.detect() works without locale segment in path');
-  } else {
-    fail(`workday.detect(no-locale) returned ${JSON.stringify(hitNoLocale)}`);
-  }
-
-  // detect() — null cases
-  if (workday.detect({ name: 'X', careers_url: 'https://example.com/careers' }) === null) {
-    pass('workday.detect() returns null for non-Workday URL');
-  } else {
-    fail('workday.detect() should return null for non-Workday URL');
-  }
-
-  // entry.api precedence: a branded careers_url is kept while the Workday tenant
-  // is pinned via api: (mirrors greenhouse/ashby).
-  const hitApiWd = workday.detect({
-    name: 'PTC',
-    careers_url: 'https://www.ptc.com/en/careers',
-    api: 'https://ptc.wd1.myworkdayjobs.com/PTC',
-  });
-  if (hitApiWd && hitApiWd.url === 'https://ptc.wd1.myworkdayjobs.com/wday/cxs/ptc/PTC/jobs') {
-    pass('workday.detect() honors api: over a branded careers_url');
-  } else {
-    fail(`workday.detect(api-pinned) returned ${JSON.stringify(hitApiWd)}`);
-  }
-
-  // A non-Workday api: must not shadow a valid Workday careers_url — resolution
-  // falls through to the next candidate instead of returning null.
-  const hitFallthrough = workday.detect({
-    name: 'Acme',
-    api: 'https://acme.com/careers',
-    careers_url: 'https://acme.wd12.myworkdayjobs.com/en-US/acme-jobs',
-  });
-  if (hitFallthrough && hitFallthrough.url === 'https://acme.wd12.myworkdayjobs.com/wday/cxs/acme/acme-jobs/jobs') {
-    pass('workday.detect() falls through a non-Workday api: to a valid careers_url');
-  } else {
-    fail(`workday.detect(fallthrough) returned ${JSON.stringify(hitFallthrough)}`);
-  }
-
-  // Path-spoofed URL: myworkdayjobs.com in path, not hostname
-  if (workday.detect({ name: 'Spoof', careers_url: 'https://evil.example/test.wd5.myworkdayjobs.com/en-US/board' }) === null) {
-    pass('workday.detect() rejects path-spoofed URL');
-  } else {
-    fail('workday.detect() must NOT detect path-spoofed URLs');
-  }
-
-  // Non-string careers_url
-  if (workday.detect({ name: 'X', careers_url: null }) === null && workday.detect({ name: 'X' }) === null) {
-    pass('workday.detect() returns null for null / missing careers_url');
-  } else {
-    fail('workday.detect() should return null for non-string careers_url');
-  }
-
-  // parseWorkdayResponse — normalization
-  const sampleJson = {
-    jobPostings: [
-      { title: 'Senior QA Engineer', externalPath: '/job/board/Senior-QA-Engineer_JR001', locationsText: 'Berlin, Germany', postedOn: 'Posted 2 Days Ago' },
-      { title: 'Lead Developer', externalPath: '/job/board/Lead-Developer_JR002', locationsText: 'Remote' },
-      { title: '', externalPath: '/job/board/No-Title_JR003' },          // no title — skip
-      { title: 'No Path Role', externalPath: '' },                        // no externalPath — skip
-      { title: 'Also No Path' },                                          // undefined externalPath — skip
-    ],
-  };
-  const entry = { name: 'Acme', careers_url: 'https://acme.wd12.myworkdayjobs.com/en-US/acme-jobs' };
-  const parsed = parseWorkdayResponse(sampleJson, entry);
-
-  if (parsed.length === 2) pass('parseWorkdayResponse extracts 2 valid jobs (skips missing title/path)');
-  else fail(`parseWorkdayResponse returned ${parsed.length} jobs, expected 2`);
-
-  if (parsed[0].title === 'Senior QA Engineer' && parsed[0].location === 'Berlin, Germany') {
-    pass('parseWorkdayResponse maps title and location');
-  } else {
-    fail(`row 0 = ${JSON.stringify(parsed[0])}`);
-  }
-
-  if (parsed[0].url.includes('acme-jobs') && parsed[0].url.includes('/job/board/Senior-QA-Engineer_JR001')) {
-    pass('parseWorkdayResponse builds URL from jobBase + externalPath');
-  } else {
-    fail(`row 0 url = ${JSON.stringify(parsed[0].url)}`);
-  }
-
-  if (parsed[0].company === 'Acme') pass('parseWorkdayResponse sets company from entry.name');
-  else fail(`parseWorkdayResponse company = ${JSON.stringify(parsed[0].company)}`);
-
-  // parseWorkdayResponse — location fallback from URL path
-  const noLocEntry = { name: 'Globex', careers_url: 'https://globex.wd103.myworkdayjobs.com/globexcareers' };
-  const noLocJson = {
-    jobPostings: [
-      { title: 'Quality Engineer', externalPath: '/job/Mumbai/Quality-Engineer_ATCI-123' },           // no locationsText
-      { title: 'Test Lead', externalPath: '/job/Remote-Poland/Test-Lead_ATCI-456', locationsText: '' }, // empty locationsText
-      { title: 'QA Analyst', externalPath: '/job/Remote-Hungary/QA-Analyst_ATCI-789', locationsText: 'Remote, Hungary' }, // has locationsText — use it
-    ],
-  };
-  const noLocParsed = parseWorkdayResponse(noLocJson, noLocEntry);
-  if (noLocParsed[0]?.location === 'Mumbai') pass('parseWorkdayResponse falls back to URL path location when locationsText absent');
-  else fail(`parseWorkdayResponse path fallback: expected "Mumbai", got ${JSON.stringify(noLocParsed[0]?.location)}`);
-  if (noLocParsed[1]?.location === 'Remote Poland') pass('parseWorkdayResponse falls back to URL path location when locationsText empty');
-  else fail(`parseWorkdayResponse path fallback empty: expected "Remote Poland", got ${JSON.stringify(noLocParsed[1]?.location)}`);
-  if (noLocParsed[2]?.location === 'Remote, Hungary') pass('parseWorkdayResponse prefers locationsText over URL path when present');
-  else fail(`parseWorkdayResponse locationsText priority: expected "Remote, Hungary", got ${JSON.stringify(noLocParsed[2]?.location)}`);
-
-  // parseWorkdayResponse — malformed percent-encoding in the URL path segment
-  // must not throw (decodeURIComponent) and must not abort processing of
-  // other job records in the same response.
-  const malformedPathJson = {
-    jobPostings: [
-      { title: 'Broken Encoding', externalPath: '/job/%E0%A4%A/Broken-Encoding_JR1' },
-      { title: 'Fine Job', externalPath: '/job/Berlin/Fine-Job_JR2' },
-    ],
-  };
-  try {
-    const malformedParsed = parseWorkdayResponse(malformedPathJson, entry);
-    if (malformedParsed.length === 2 && malformedParsed[1].location === 'Berlin') {
-      pass('parseWorkdayResponse tolerates malformed percent-encoding without dropping other records');
-    } else {
-      fail(`parseWorkdayResponse malformed encoding result = ${JSON.stringify(malformedParsed)}`);
-    }
-  } catch (e4) {
-    fail(`parseWorkdayResponse should not throw on malformed percent-encoding: ${e4.message}`);
-  }
-
-  // parseWorkdayResponse — empty / malformed input
-  if (parseWorkdayResponse({}, entry).length === 0) pass('parseWorkdayResponse({}) → empty result');
-  else fail('parseWorkdayResponse({}) should be empty');
-
-  if (parseWorkdayResponse({ jobPostings: null }, entry).length === 0) {
-    pass('parseWorkdayResponse handles null jobPostings');
-  } else {
-    fail('parseWorkdayResponse null jobPostings should be empty');
-  }
-
-  // fetch() with mock ctx — uses total field to bound sequential pagination
-  let postRequests = 0;
-  const capturedRedirects = [];
-  const fetchedJobs = await workday.fetch(entry, mkWorkdayCtx(async (_url, opts) => {
-    postRequests++;
-    capturedRedirects.push(opts?.redirect);
-    const body = JSON.parse(opts.body);
-    if (body.offset === 0) {
-      return { total: 30, jobPostings: Array.from({ length: 20 }, (_, i) => ({ title: `Job P1-${i}`, externalPath: `/job/board/p1-${i}` })) };
-    }
-    return { total: 30, jobPostings: Array.from({ length: 10 }, (_, i) => ({ title: `Job P2-${i}`, externalPath: `/job/board/p2-${i}` })) };
-  }));
-  if (postRequests === 2 && fetchedJobs.length === 30) {
-    pass('workday.fetch() uses total field to fetch exact pages sequentially (20+10=30)');
-  } else {
-    fail(`fetch pagination: requests=${postRequests}, total=${fetchedJobs.length} (expected 2 requests / 30 jobs)`);
-  }
-
-  if (capturedRedirects.length === 2 && capturedRedirects.every(r => r === 'error')) {
-    pass('workday.fetch() passes redirect:"error" on every page (SSRF guard)');
-  } else {
-    fail(`workday.fetch() redirect opts across pages = ${JSON.stringify(capturedRedirects)}`);
-  }
-
-  // parseWorkdayResponse — null/undefined entries in jobPostings must be
-  // skipped, not crash
-  const sparseWorkday = { jobPostings: [null, undefined, { title: 'Real Job', externalPath: '/job/board/real-job' }] };
-  try {
-    const parsedSparseWorkday = parseWorkdayResponse(sparseWorkday, entry);
-    if (parsedSparseWorkday.length === 1 && parsedSparseWorkday[0].title === 'Real Job') {
-      pass('parseWorkdayResponse skips null/undefined entries without crashing');
-    } else {
-      fail(`parseWorkdayResponse sparse result = ${JSON.stringify(parsedSparseWorkday)}`);
-    }
-  } catch (e2) {
-    fail(`parseWorkdayResponse should not throw on null/undefined entries: ${e2.message}`);
-  }
-
-  // fetch() pagination cap — an inflated `total` must not trigger unbounded
-  // requests (DEFAULT_MAX_PAGES = 100 in providers/workday.mjs), and hitting
-  // the cap must be visible (console.error), not silent — real tenants
-  // (Dollar Tree, total=23,609; CVS Health, total=16,974) already exceed the
-  // 100-page/2000-job default.
-  let hugeWorkdayRequests = 0;
-  const capturedWarnings = [];
-  const originalConsoleError = console.error;
-  console.error = (msg) => capturedWarnings.push(msg);
-  let fetchedHugeWorkday;
-  try {
-    fetchedHugeWorkday = await workday.fetch(entry, mkWorkdayCtx(async () => {
-      hugeWorkdayRequests++;
-      return { total: 1_000_000, jobPostings: Array.from({ length: 20 }, (_, i) => ({ title: `Job ${i}`, externalPath: `/job/board/${i}` })) };
-    }));
-  } finally {
-    console.error = originalConsoleError;
-  }
-  if (hugeWorkdayRequests === 100 && fetchedHugeWorkday.length === 2000) {
-    pass('workday.fetch() caps pagination at DEFAULT_MAX_PAGES despite an inflated total');
-  } else {
-    fail(`workday fetch pagination cap: requests=${hugeWorkdayRequests}, total=${fetchedHugeWorkday.length} (expected 100/2000)`);
-  }
-  if (capturedWarnings.some(w => /truncated at max_pages=\d+/.test(w))) {
-    pass('workday.fetch() warns (console.error) when the cap truncates real results');
-  } else {
-    fail(`workday fetch cap: expected a truncation warning, got ${JSON.stringify(capturedWarnings)}`);
-  }
-
-  // fetch() pagination cap — entry.max_pages raises the cap for a genuinely
-  // large tenant (e.g. Deutsche Bank-scale postings)
-  let overriddenWorkdayRequests = 0;
-  const bigWorkdayEntry = { name: 'BigCo', careers_url: 'https://bigco.wd5.myworkdayjobs.com/careers', max_pages: 80 };
-  const fetchedOverriddenWorkday = await workday.fetch(bigWorkdayEntry, mkWorkdayCtx(async () => {
-    overriddenWorkdayRequests++;
-    return { total: 1_000_000, jobPostings: Array.from({ length: 20 }, (_, i) => ({ title: `Job ${i}`, externalPath: `/job/board/${i}` })) };
-  }));
-  if (overriddenWorkdayRequests === 80 && fetchedOverriddenWorkday.length === 1600) {
-    pass('workday.fetch() honors entry.max_pages to raise the cap above the default');
-  } else {
-    fail(`workday fetch max_pages override: requests=${overriddenWorkdayRequests}, total=${fetchedOverriddenWorkday.length} (expected 80/1600)`);
-  }
-
-  // entry.max_pages is itself capped (MAX_PAGES_CAP = 1500) — an absurd
-  // override can't turn this into an unbounded scan either.
-  let absurdWorkdayRequests = 0;
-  const absurdWorkdayEntry = { name: 'AbsurdCo', careers_url: 'https://absurdco.wd5.myworkdayjobs.com/careers', max_pages: 100_000 };
-  const fetchedAbsurdWorkday = await workday.fetch(absurdWorkdayEntry, mkWorkdayCtx(async () => {
-    absurdWorkdayRequests++;
-    return { total: 10_000_000, jobPostings: Array.from({ length: 20 }, (_, i) => ({ title: `Job ${i}`, externalPath: `/job/board/${i}` })) };
-  }));
-  if (absurdWorkdayRequests === 1500 && fetchedAbsurdWorkday.length === 30_000) {
-    pass('workday.fetch() caps an absurd entry.max_pages at MAX_PAGES_CAP');
-  } else {
-    fail(`workday fetch max_pages hard cap: requests=${absurdWorkdayRequests}, total=${fetchedAbsurdWorkday.length} (expected 1500/30000)`);
-  }
-
-  // Invalid max_pages values (negative, zero, non-numeric) fall back to
-  // DEFAULT_MAX_PAGES, same as omitting max_pages entirely.
-  for (const invalidMaxPages of [-5, 0, 'abc', NaN, null]) {
-    let invalidRequests = 0;
-    const invalidEntry = { name: 'InvalidCo', careers_url: 'https://invalidco.wd5.myworkdayjobs.com/careers', max_pages: invalidMaxPages };
-    const fetchedInvalid = await workday.fetch(invalidEntry, mkWorkdayCtx(async () => {
-      invalidRequests++;
-      return { total: 1_000_000, jobPostings: Array.from({ length: 20 }, (_, i) => ({ title: `Job ${i}`, externalPath: `/job/board/${i}` })) };
-    }));
-    // Number.isNaN(NaN) is true but JSON.stringify(NaN) === 'null', which would
-    // be indistinguishable from the literal `null` case in the log below.
-    const label = Number.isNaN(invalidMaxPages) ? 'NaN' : JSON.stringify(invalidMaxPages);
-    if (invalidRequests === 100 && fetchedInvalid.length === 2000) {
-      pass(`workday.fetch() falls back to DEFAULT_MAX_PAGES for invalid max_pages=${label}`);
-    } else {
-      fail(`workday fetch invalid max_pages=${label}: requests=${invalidRequests}, total=${fetchedInvalid.length} (expected 100/2000)`);
-    }
-  }
-
-  // fetch() pagination — a failure that persists across every retry attempt
-  // on a later page returns the jobs gathered so far instead of discarding
-  // everything (sequential, not Promise.all), retries MAX_RETRIES+1=4 times
-  // on that page before giving up, and the failure itself is visible
-  // (console.error), not silent. The truncation ("raise max_pages") warning
-  // must NOT also fire — that knob does nothing for a rate-limited tenant.
-  let flakyWorkdayRequests = 0;
-  const flakyWarnings = [];
-  console.error = (msg) => flakyWarnings.push(msg);
-  let flakyWorkdayJobs;
-  try {
-    flakyWorkdayJobs = await workday.fetch(entry, mkWorkdayCtx(async (_url, opts) => {
-      flakyWorkdayRequests++;
-      const body = JSON.parse(opts.body);
-      const page = body.offset / 20; // PAGE_SIZE in providers/workday.mjs
-      if (page === 2) { const err = new Error('HTTP 503'); err.status = 503; throw err; }
-      return { total: 80, jobPostings: Array.from({ length: 20 }, (_, i) => ({ title: `Job p${page}-${i}`, externalPath: `/job/board/p${page}-${i}` })) };
-    }));
-  } finally {
-    console.error = originalConsoleError;
-  }
-  if (flakyWorkdayRequests === 6 && flakyWorkdayJobs.length === 40) {
-    pass('workday.fetch() retries a failing page 4x then returns partial results');
-  } else {
-    fail(`workday fetch partial failure: requests=${flakyWorkdayRequests}, total=${flakyWorkdayJobs.length} (expected 6/40)`);
-  }
-  if (flakyWarnings.some(w => /truncated at \d+ of \d+ pages after 4 attempts/.test(w))) {
-    pass('workday.fetch() warns (console.error) when a page fetch fails after exhausting retries, with the attempt count');
-  } else {
-    fail(`workday fetch page failure: expected a fetch-failed warning, got ${JSON.stringify(flakyWarnings)}`);
-  }
-  // The failure message must show scale — which page out of how many were
-  // planned, and how many jobs came back out of the tenant's real total —
-  // not just "page 3, 40 jobs" with no sense of how much was actually lost.
-  // Same "truncated at ... (N of M jobs)" shape as the cap-hit warning below,
-  // so the two read consistently.
-  if (flakyWarnings.some(w => /truncated at 3 of 4 pages after 4 attempts \(40 of 80 jobs\): HTTP 503/.test(w))) {
-    pass('workday.fetch() fetch-failure warning reports scale (page X of Y, N of total jobs)');
-  } else {
-    fail(`workday fetch-failure warning missing scale context: ${JSON.stringify(flakyWarnings)}`);
-  }
-  if (!flakyWarnings.some(w => /raise max_pages/.test(w))) {
-    pass('workday.fetch() does NOT fire the "raise max_pages" warning on a fetch-error stop');
-  } else {
-    fail(`workday fetch-error stop should not also warn about max_pages: ${JSON.stringify(flakyWarnings)}`);
-  }
-
-  // fetch() retry — a 429 that succeeds on a later attempt is transparent to
-  // the caller (no jobs lost, no error surfaced) and respects Retry-After.
-  let retrySleepCalls = [];
-  let retryAttempts = 0;
-  const retryEntry = { name: 'RetryCo', careers_url: 'https://retryco.wd5.myworkdayjobs.com/careers' };
-  const retryJobs = await workday.fetch(retryEntry, mkWorkdayCtx(async () => {
-    retryAttempts++;
-    if (retryAttempts === 1) { const err = new Error('HTTP 429'); err.status = 429; err.retryAfter = '1'; throw err; }
-    return { total: 1, jobPostings: [{ title: 'Recovered Job', externalPath: '/job/board/recovered' }] };
-  }, { sleep: async (ms) => { retrySleepCalls.push(ms); } }));
-  if (retryAttempts === 2 && retryJobs.length === 1 && retryJobs[0].title === 'Recovered Job') {
-    pass('workday.fetch() retries a 429 and recovers transparently');
-  } else {
-    fail(`workday 429 retry: attempts=${retryAttempts}, jobs=${JSON.stringify(retryJobs)}`);
-  }
-  if (retrySleepCalls[0] === 1000) {
-    pass('workday.fetch() honors Retry-After header for backoff delay');
-  } else {
-    fail(`workday retry-after: expected first backoff delay 1000ms, got ${JSON.stringify(retrySleepCalls)}`);
-  }
-
-  // fetch() retry — a hostile or misconfigured Retry-After (e.g. 86400s = a
-  // full day) must not be honored verbatim: it's clamped to
-  // RETRY_MAX_DELAY_MS * 4 (32s) so a single bad header can't stall a
-  // tenant's fetch indefinitely, defeating the point of a bounded backoff.
-  let hostileRetrySleepCalls = [];
-  let hostileRetryAttempts = 0;
-  const hostileRetryEntry = { name: 'HostileRetryCo', careers_url: 'https://hostileretryco.wd5.myworkdayjobs.com/careers' };
-  await workday.fetch(hostileRetryEntry, mkWorkdayCtx(async () => {
-    hostileRetryAttempts++;
-    if (hostileRetryAttempts === 1) { const err = new Error('HTTP 429'); err.status = 429; err.retryAfter = '86400'; throw err; }
-    return { total: 0, jobPostings: [] };
-  }, { sleep: async (ms) => { hostileRetrySleepCalls.push(ms); } }));
-  if (hostileRetrySleepCalls[0] === 32_000) {
-    pass('workday.fetch() clamps an oversized Retry-After to RETRY_MAX_DELAY_MS * 4');
-  } else {
-    fail(`workday retry-after clamp: expected 32000ms, got ${JSON.stringify(hostileRetrySleepCalls)}`);
-  }
-
-  // fetch() retry — a non-retryable 4xx (e.g. malformed request) breaks
-  // immediately, without wasting retry attempts.
-  let non429Attempts = 0;
-  const non429Warnings = [];
-  console.error = (msg) => non429Warnings.push(msg);
-  let non429Jobs;
-  try {
-    non429Jobs = await workday.fetch(entry, mkWorkdayCtx(async (_url, opts) => {
-      non429Attempts++;
-      const body = JSON.parse(opts.body);
-      if (body.offset === 0) return { total: 40, jobPostings: Array.from({ length: 20 }, (_, i) => ({ title: `Job ${i}`, externalPath: `/job/board/${i}` })) };
-      const err = new Error('HTTP 400: bad request'); err.status = 400; throw err;
-    }));
-  } finally {
-    console.error = originalConsoleError;
-  }
-  if (non429Attempts === 2 && non429Jobs.length === 20) {
-    pass('workday.fetch() does not retry a non-retryable 4xx error');
-  } else {
-    fail(`workday non-retryable 4xx: attempts=${non429Attempts}, jobs=${non429Jobs.length} (expected 2/20)`);
-  }
-
-  // fetch() early-stop — once a page's postings are all clearly past
-  // ctx.sinceMs, pagination stops without hitting max_pages, and the
-  // "raise max_pages" warning does NOT fire (this isn't a cap hit).
-  const SINCE_DAYS = 3; // mirrors scan-ats-full.mjs's --since default
-  const nowMs = Date.now();
-  const sinceMs = nowMs - SINCE_DAYS * 86_400_000;
-  let earlyStopRequests = 0;
-  const earlyStopWarnings = [];
-  console.error = (msg) => earlyStopWarnings.push(msg);
-  let earlyStopJobs;
-  try {
-    earlyStopJobs = await workday.fetch(entry, mkWorkdayCtx(async (_url, opts) => {
-      earlyStopRequests++;
-      const body = JSON.parse(opts.body);
-      const page = body.offset / 20;
-      if (page === 0) {
-        // Page 0: fresh postings, well within the window.
-        return { total: 1_000_000, jobPostings: Array.from({ length: 20 }, (_, i) => ({ title: `Fresh ${i}`, externalPath: `/job/board/fresh-${i}`, postedOn: 'Posted Today' })) };
-      }
-      // Every later page: clearly stale (well past sinceMs - margin) — if
-      // early-stop didn't work, this mock would be asked for 100 pages.
-      return { total: 1_000_000, jobPostings: Array.from({ length: 20 }, (_, i) => ({ title: `Stale ${i}`, externalPath: `/job/board/stale-${i}`, postedOn: 'Posted 20 Days Ago' })) };
-    }, { sinceMs }));
-  } finally {
-    console.error = originalConsoleError;
-  }
-  if (earlyStopRequests === 2 && earlyStopJobs.length === 40) {
-    pass('workday.fetch() stops paginating once a page is past --since (early-stop)');
-  } else {
-    fail(`workday early-stop: requests=${earlyStopRequests}, jobs=${earlyStopJobs.length} (expected 2/40)`);
-  }
-  if (!earlyStopWarnings.some(w => /truncated at/.test(w))) {
-    pass('workday.fetch() does NOT warn about max_pages when it stopped early on --since');
-  } else {
-    fail(`workday early-stop should not warn about max_pages: ${JSON.stringify(earlyStopWarnings)}`);
-  }
-
-  // fetch() early-stop — a wide --since window (>= 30 days) never triggers
-  // early-stop off the unbounded "30+ Days Ago" bucket; pagination still
-  // proceeds until max_pages/total, as before. includeUndated: true isolates
-  // this from the no-date-skip optimization tested separately below — every
-  // posting here is undated ("30+"), which would otherwise short-circuit
-  // after page 1 regardless of --since width.
-  const WIDE_SINCE_DAYS = 90; // >= 30 — past the "30+ Days Ago" bucket's ambiguity threshold
-  const wideSinceMs = nowMs - WIDE_SINCE_DAYS * 86_400_000;
-  let wideRequests = 0;
-  const wideJobs = await workday.fetch(entry, mkWorkdayCtx(async (_url, opts) => {
-    wideRequests++;
-    const body = JSON.parse(opts.body);
-    if (body.offset === 0) {
-      return { total: 40, jobPostings: Array.from({ length: 20 }, (_, i) => ({ title: `Old ${i}`, externalPath: `/job/board/old-${i}`, postedOn: 'Posted 30+ Days Ago' })) };
-    }
-    return { total: 40, jobPostings: Array.from({ length: 20 }, (_, i) => ({ title: `Old2 ${i}`, externalPath: `/job/board/old2-${i}`, postedOn: 'Posted 30+ Days Ago' })) };
-  }, { sinceMs: wideSinceMs, includeUndated: true }));
-  if (wideRequests === 2 && wideJobs.length === 40) {
-    pass('workday.fetch() never early-stops off the unbounded "30+ Days Ago" bucket');
-  } else {
-    fail(`workday wide-since: requests=${wideRequests}, jobs=${wideJobs.length} (expected 2/40)`);
-  }
-
-  // fetch() cap-hit warning — reverse-scan context (ctx.sinceMs set, as
-  // scan-ats-full.mjs always does) where entries are synthesized from an
-  // external dataset, not portals.yml: there's no portal entry to edit, and
-  // — per the "no fixed cap can guarantee full coverage" conclusion — no
-  // fix to advise at all, so the message is just the short fact, with
-  // neither "raise max_pages" nor a portals.yml edit suggested.
-  // includeUndated: true forces this past the no-date-skip short-circuit
-  // (tested separately below) so the fetch actually reaches the cap.
-  const noDateSinceMs = nowMs - SINCE_DAYS * 86_400_000;
-  const noDateWarnings = [];
-  console.error = (msg) => noDateWarnings.push(msg);
-  try {
-    await workday.fetch(entry, mkWorkdayCtx(async () => ({
-      total: 1_000_000,
-      jobPostings: Array.from({ length: 20 }, (_, i) => ({ title: `NoDate ${i}`, externalPath: `/job/board/nodate-${i}` })), // no postedOn
-    }), { sinceMs: noDateSinceMs, includeUndated: true }));
-  } finally {
-    console.error = originalConsoleError;
-  }
-  if (noDateWarnings.some(w => /truncated at \d+ pages/.test(w))) {
-    pass('workday.fetch() cap-hit warning fires in reverse-scan context (tenant has no dates, includeUndated on)');
-  } else {
-    fail(`workday no-date cap warning missing: ${JSON.stringify(noDateWarnings)}`);
-  }
-  if (!noDateWarnings.some(w => /raise max_pages|portal entry|portals\.yml/.test(w))) {
-    pass('workday.fetch() reverse-scan cap warning gives no inactionable advice (no portal entry to edit)');
-  } else {
-    fail(`workday reverse-scan cap warning should not suggest editing a portal entry: ${JSON.stringify(noDateWarnings)}`);
-  }
-
-  // fetch() no-date-skip — the default case (includeUndated NOT set, as
-  // scan-ats-full.mjs leaves it by default): a tenant whose first page has
-  // zero dated postings stops right there instead of grinding through up to
-  // maxPages requests whose results would all be dropped as 'undated'
-  // downstream anyway. Only 1 request should fire, not maxPages (100).
-  const skipSinceMs = nowMs - SINCE_DAYS * 86_400_000;
-  let skipRequests = 0;
-  const skipWarnings = [];
-  console.error = (msg) => skipWarnings.push(msg);
-  let skipJobs;
-  try {
-    skipJobs = await workday.fetch(entry, mkWorkdayCtx(async () => {
-      skipRequests++;
-      return {
-        total: 1_000_000,
-        jobPostings: Array.from({ length: 20 }, (_, i) => ({ title: `NoDate ${i}`, externalPath: `/job/board/skip-${i}` })), // no postedOn
-      };
-    }, { sinceMs: skipSinceMs })); // includeUndated intentionally omitted — the default, falsy case
-  } finally {
-    console.error = originalConsoleError;
-  }
-  if (skipRequests === 1 && skipJobs.length === 20) {
-    pass('workday.fetch() skips pagination after page 1 when includeUndated is off and the tenant has no dated postings');
-  } else {
-    fail(`workday no-date-skip: requests=${skipRequests}, jobs=${skipJobs.length} (expected 1/20)`);
-  }
-  // A full-directory scan hits this on a large fraction of tenants — a
-  // console.error per occurrence would just be the same line thousands of
-  // times, so the signal is a tag on the returned array instead (aggregated
-  // by scan-ats-full.mjs into one summary line at the end of the run).
-  if (skipJobs.workdayNoDateSkip === true) {
-    pass('workday.fetch() tags the returned jobs array for no-date-skip aggregation');
-  } else {
-    fail(`workday no-date-skip tag missing: jobs.workdayNoDateSkip = ${JSON.stringify(skipJobs.workdayNoDateSkip)}`);
-  }
-  if (skipWarnings.length === 0) {
-    pass('workday.fetch() does not console.error per-company on a no-date-skip (aggregated instead)');
-  } else {
-    fail(`workday no-date-skip should not log anything directly: ${JSON.stringify(skipWarnings)}`);
-  }
-
-  // fetch() no-date-skip — does NOT trigger when includeUndated is true (the
-  // "hit the cap while genuinely undated" scenario above already covers that
-  // the fetch continues in that case); this just re-confirms the gate.
-  let noSkipWithIncludeUndatedRequests = 0;
-  const noSkipJobs = await workday.fetch(entry, mkWorkdayCtx(async (_url, opts) => {
-    noSkipWithIncludeUndatedRequests++;
-    const body = JSON.parse(opts.body);
-    if (body.offset === 0) return { total: 40, jobPostings: Array.from({ length: 20 }, (_, i) => ({ title: `ND ${i}`, externalPath: `/job/board/nd-${i}` })) };
-    return { total: 40, jobPostings: Array.from({ length: 20 }, (_, i) => ({ title: `ND2 ${i}`, externalPath: `/job/board/nd2-${i}` })) };
-  }, { sinceMs: skipSinceMs, includeUndated: true }));
-  if (noSkipWithIncludeUndatedRequests === 2 && noSkipJobs.length === 40) {
-    pass('workday.fetch() does not no-date-skip when includeUndated is true');
-  } else {
-    fail(`workday no-date-skip gate: requests=${noSkipWithIncludeUndatedRequests}, jobs=${noSkipJobs.length} (expected 2/40)`);
-  }
-
-  // fetch() cap-hit warning — reverse-scan context, tenant genuinely has
-  // more within --since than the cap allows (total far above the window,
-  // e.g. cvshealth-scale): short line, no suspect-cap tag.
-  const datedCapSinceMs = nowMs - SINCE_DAYS * 86_400_000;
-  const datedCapWarnings = [];
-  console.error = (msg) => datedCapWarnings.push(msg);
-  try {
-    await workday.fetch(entry, mkWorkdayCtx(async () => ({
-      total: 1_000_000,
-      jobPostings: Array.from({ length: 20 }, (_, i) => ({ title: `Fresh ${i}`, externalPath: `/job/board/fresh-${i}`, postedOn: 'Posted Today' })),
-    }), { sinceMs: datedCapSinceMs }));
-  } finally {
-    console.error = originalConsoleError;
-  }
-  if (datedCapWarnings.some(w => /truncated at \d+ pages \(2000 of 1000000 jobs\)/.test(w))) {
-    pass('workday.fetch() cap-hit warning reports the short "truncated at N pages" form');
-  } else {
-    fail(`workday dated cap-hit warning mismatch: ${JSON.stringify(datedCapWarnings)}`);
-  }
-  if (!datedCapWarnings.some(w => /Workday-capped/.test(w))) {
-    pass('workday.fetch() cap-hit warning omits the suspected-Workday-cap tag when total is far above the window');
-  } else {
-    fail(`workday cap-hit warning should not suspect a Workday-side cap here (total=1,000,000, window=2,000): ${JSON.stringify(datedCapWarnings)}`);
-  }
-
-  // fetch() cap-hit warning — Workday's own CXS backend has been observed
-  // reporting `total` as exactly maxPages*PAGE_SIZE even when the real count
-  // is far higher (verified live: dickssportinggoods reported total=2000 but
-  // its public careers page listed 7,120 openings, and offset=2000/4000
-  // requests returned the same first posting as offset=0 instead of new
-  // results). This exact-match case must carry a distinct, short tag.
-  const suspectCapSinceMs = nowMs - SINCE_DAYS * 86_400_000;
-  const suspectCapWarnings = [];
-  console.error = (msg) => suspectCapWarnings.push(msg);
-  try {
-    await workday.fetch(entry, mkWorkdayCtx(async () => ({
-      total: 2000, // === DEFAULT_MAX_PAGES (100) * PAGE_SIZE (20)
-      jobPostings: Array.from({ length: 20 }, (_, i) => ({ title: `Suspect ${i}`, externalPath: `/job/board/suspect-${i}`, postedOn: 'Posted Today' })),
-    }), { sinceMs: suspectCapSinceMs }));
-  } finally {
-    console.error = originalConsoleError;
-  }
-  if (suspectCapWarnings.some(w => /\(total may be Workday-capped, not real\)/.test(w))) {
-    pass('workday.fetch() cap-hit warning flags a suspected Workday-side total cap when total === maxPages*PAGE_SIZE');
-  } else {
-    fail(`workday suspected-cap tag missing: ${JSON.stringify(suspectCapWarnings)}`);
-  }
-
-  // Verify POST method was used
-  let capturedMethod = null;
-  await workday.fetch(entry, mkWorkdayCtx(async (_url, opts) => { capturedMethod = opts?.method; return { total: 0, jobPostings: [] }; }));
-  if (capturedMethod === 'POST') pass('workday.fetch() uses POST method');
-  else fail(`workday.fetch() method is ${JSON.stringify(capturedMethod)}, expected POST`);
-
-  // Fallback: no total field — paginate sequentially until short page
-  let fallbackRequests = 0;
-  const fallbackJobs = await workday.fetch(entry, mkWorkdayCtx(async (_url, opts) => {
-    fallbackRequests++;
-    const body = JSON.parse(opts.body);
-    if (body.offset === 0) return { jobPostings: Array.from({ length: 20 }, (_, i) => ({ title: `FB P1-${i}`, externalPath: `/job/board/fb-${i}` })) };
-    return { jobPostings: Array.from({ length: 5 }, (_, i) => ({ title: `FB P2-${i}`, externalPath: `/job/board/fb2-${i}` })) };
-  }));
-  if (fallbackRequests === 2 && fallbackJobs.length === 25) {
-    pass('workday.fetch() fallback (no total): paginates sequentially, stops on short page (20+5=25)');
-  } else {
-    fail(`fallback pagination: requests=${fallbackRequests}, jobs=${fallbackJobs.length} (expected 2/25)`);
-  }
-
-  // fetch() honors ctx.maxPages — verify-portals' liveness probe sets maxPages:1.
-  // It must stop after the first page and NOT request page 2, which would trip
-  // the probe's second-request sentinel; fetchPageWithRetry treats that abort as
-  // transient and retries it MAX_RETRIES times (4 requests) plus a truncation
-  // warning. Even though total=173 implies 9 pages, the cap must win at 1.
-  let probeRequests = 0;
-  const probeWarnings = [];
-  const probeErr = console.error;
-  console.error = (m) => probeWarnings.push(m);
-  let probeJobs;
-  try {
-    probeJobs = await workday.fetch(entry, mkWorkdayCtx(async (_url, opts) => {
-      probeRequests++;
-      // Simulate the probe sentinel: any request past the first throws.
-      if (JSON.parse(opts.body).offset > 0) throw new Error('probe budget: no second page');
-      return { total: 173, jobPostings: Array.from({ length: 20 }, (_, i) => ({ title: `Job ${i}`, externalPath: `/job/board/${i}` })) };
-    }, { maxPages: 1 }));
-  } finally {
-    console.error = probeErr;
-  }
-  if (probeRequests === 1 && probeJobs.length === 20) {
-    pass('workday.fetch() honors ctx.maxPages=1 — one request, no second-page retry storm');
-  } else {
-    fail(`workday probe cap: requests=${probeRequests} (expected 1), jobs=${probeJobs?.length} (expected 20)`);
-  }
-  if (!probeWarnings.some((w) => /truncated/.test(String(w)))) {
-    pass('workday.fetch() stays silent (no truncation warning) under the liveness probe cap');
-  } else {
-    fail(`workday probe should emit no warning, got: ${JSON.stringify(probeWarnings)}`);
-  }
-
-} catch (e) {
-  fail(`workday provider tests crashed: ${e.message}`);
-}
-
 // ── 54. _http.mjs — error messages are status code + reason phrase only ──
 // WAF challenge pages (seen live: Workday 429s) carry no actionable text —
 // whether it's raw HTML markup or a human-readable challenge page ("Security
@@ -6575,371 +5780,6 @@ try {
   }
 } catch (e) {
   fail(`_http.mjs error message tests crashed: ${e.message}`);
-}
-
-// ── 52. Provider — getonbrd ─────────────────────────────────────
-console.log('\n52. Provider — getonbrd');
-
-try {
-  const getonbrdModule = await import(pathToFileURL(join(ROOT, 'providers/getonbrd.mjs')).href);
-  const getonbrd = getonbrdModule.default;
-
-  if (getonbrd.id === 'getonbrd') pass('getonbrd.id is "getonbrd"');
-  else fail(`getonbrd.id is ${JSON.stringify(getonbrd.id)}`);
-
-  // Deterministic JSON:API sample — no network. Two valid jobs plus two dropped
-  // (empty title, non-absolute url).
-  const sample = {
-    data: [
-      {
-        attributes: {
-          title: 'Staff AI Engineer',
-          remote: true,
-          countries: 'Remote',
-          company: { data: { attributes: { name: 'Acme Corp' } } },
-        },
-        links: { public_url: 'https://www.getonbrd.com/jobs/acme-staff-ai-engineer' },
-      },
-      {
-        attributes: {
-          title: '  Platform Engineer  ',                  // leading/trailing space → trimmed
-          remote: false,
-          countries: ['Chile'],                            // live API sends an array of country names
-          published_at: 1700000000,                        // epoch seconds → postedAt in ms
-          company: { data: { attributes: { name: '' } } }, // empty → falls back to entry.name
-        },
-        links: { public_url: '  https://www.getonbrd.com/jobs/beta-platform-engineer  ' },
-      },
-      {
-        attributes: { title: '', company: { data: { attributes: { name: 'Bad Co' } } } }, // dropped: empty title
-        links: { public_url: 'https://www.getonbrd.com/jobs/bad-empty-title' },
-      },
-      {
-        attributes: { title: 'Relative URL Role', remote: true },                          // dropped: non-absolute url
-        links: { public_url: '/jobs/relative' },
-      },
-    ],
-  };
-
-  let capturedUrl = null;
-  let capturedOpts = null;
-  const fetched = await getonbrd.fetch(
-    { name: 'GetOnBoard Feed', provider: 'getonbrd' },
-    { fetchJson: async (url, opts) => { capturedUrl = url; capturedOpts = opts; return sample; } },
-  );
-
-  if (capturedUrl === 'https://www.getonbrd.com/api/v0/categories/programming/jobs?per_page=100&expand[]=company&page=1')
-    pass('getonbrd.fetch() requests the board-wide category feed URL (page 1)');
-  else fail(`getonbrd.fetch() requested ${JSON.stringify(capturedUrl)}`);
-
-  if (capturedOpts && capturedOpts.redirect === 'error')
-    pass('getonbrd.fetch() passes redirect:"error" to fetchJson (SSRF guard)');
-  else fail(`getonbrd.fetch() should pass redirect:"error", got: ${JSON.stringify(capturedOpts)}`);
-
-  if (fetched.length === 2)
-    pass('getonbrd.fetch() keeps 2 valid jobs (drops empty-title + non-absolute-url rows)');
-  else fail(`getonbrd.fetch() returned ${fetched.length} jobs (expected 2)`);
-
-  if (fetched[0] && Object.keys(fetched[0]).sort().join(',') === 'company,location,title,url')
-    pass('getonbrd.fetch() returns the normalized { title, url, company, location } shape');
-  else fail(`getonbrd.fetch() row 0 keys = ${JSON.stringify(fetched[0] && Object.keys(fetched[0]))}`);
-
-  if (fetched[0]?.title === 'Staff AI Engineer'
-      && fetched[0]?.url === 'https://www.getonbrd.com/jobs/acme-staff-ai-engineer'
-      && fetched[0]?.company === 'Acme Corp'
-      && fetched[0]?.location === 'Remote')
-    pass('getonbrd.fetch() maps title/url/company and uses "Remote" when remote===true');
-  else fail(`getonbrd.fetch() row 0 = ${JSON.stringify(fetched[0])}`);
-
-  if (fetched[1]?.title === 'Platform Engineer'
-      && fetched[1]?.url === 'https://www.getonbrd.com/jobs/beta-platform-engineer')
-    pass('getonbrd.fetch() trims whitespace from title and url');
-  else fail(`getonbrd.fetch() row 1 title/url = ${JSON.stringify({ title: fetched[1]?.title, url: fetched[1]?.url })}`);
-
-  if (fetched[1]?.company === 'GetOnBoard Feed')
-    pass('getonbrd.fetch() falls back to entry.name when the company name is empty');
-  else fail(`getonbrd.fetch() row 1 company = ${JSON.stringify(fetched[1]?.company)}`);
-
-  if (fetched[1]?.location === 'Chile')
-    pass('getonbrd.fetch() joins the countries array into location when not remote');
-  else fail(`getonbrd.fetch() row 1 location = ${JSON.stringify(fetched[1]?.location)}`);
-
-  if (fetched[1]?.postedAt === 1700000000 * 1000)
-    pass('getonbrd.fetch() maps published_at (epoch seconds) to postedAt in ms');
-  else fail(`getonbrd.fetch() row 1 postedAt = ${JSON.stringify(fetched[1]?.postedAt)}`);
-
-  const noName = await getonbrd.fetch(
-    {},
-    { fetchJson: async () => ({ data: [{ attributes: { title: 'Role', remote: true }, links: { public_url: 'https://www.getonbrd.com/jobs/x' } }] }) },
-  );
-  if (noName[0]?.company === 'Get on Board')
-    pass('getonbrd.fetch() defaults company to "Get on Board" when name and entry.name are both missing');
-  else fail(`getonbrd.fetch() default company = ${JSON.stringify(noName[0]?.company)}`);
-
-  let badResponseThrew = false;
-  try {
-    await getonbrd.fetch(
-      { name: 'X', provider: 'getonbrd' },
-      { fetchJson: async () => ({ wrong: true }) },
-    );
-  } catch (e) {
-    badResponseThrew = /unexpected API response/.test(e.message);
-  }
-  if (badResponseThrew) pass('getonbrd.fetch() throws on unexpected API response shape');
-  else fail('getonbrd.fetch() should throw when the data array is absent');
-
-  // Pagination: a full first page (=per_page) is followed until a short page.
-  const mkJob = i => ({ attributes: { title: `Role ${i}`, remote: true }, links: { public_url: `https://www.getonbrd.com/jobs/role-${i}` } });
-  const fullPage = { data: Array.from({ length: 100 }, (_, i) => mkJob(i)) };
-  const shortPage = { data: [mkJob(999)] };
-
-  const pageCalls = [];
-  const paged = await getonbrd.fetch(
-    { name: 'GOB', provider: 'getonbrd' },
-    { fetchJson: async (url, opts) => { pageCalls.push({ url, opts }); return pageCalls.length === 1 ? fullPage : shortPage; } },
-  );
-  const pageUrls = pageCalls.map((c) => c.url);
-  if (pageCalls.length === 2 && /[?&]page=1(?:&|$)/.test(pageUrls[0]) && /[?&]page=2(?:&|$)/.test(pageUrls[1]))
-    pass('getonbrd.fetch() paginates ?page=N until a short page is returned');
-  else fail(`getonbrd.fetch() page URLs = ${JSON.stringify(pageUrls)}`);
-
-  if (pageCalls.length > 1 && pageCalls.every((c) => c.opts && c.opts.redirect === 'error'))
-    pass('getonbrd.fetch() passes redirect:"error" on every paginated request (not just page 1)');
-  else fail(`getonbrd.fetch() paginated opts = ${JSON.stringify(pageCalls.map((c) => c.opts))}`);
-
-  if (paged.length === 101)
-    pass('getonbrd.fetch() accumulates jobs across pages (100 + 1)');
-  else fail(`getonbrd.fetch() paginated total = ${paged.length} (expected 101)`);
-
-  const capCalls = [];
-  await getonbrd.fetch(
-    { name: 'GOB', max_pages: 1 },
-    { fetchJson: async (url) => { capCalls.push(url); return fullPage; } },
-  );
-  if (capCalls.length === 1)
-    pass('getonbrd.fetch() respects the max_pages override (stops after 1 page)');
-  else fail(`getonbrd.fetch() max_pages=1 made ${capCalls.length} page calls`);
-
-} catch (e) {
-  fail(`getonbrd provider tests crashed: ${e.message}`);
-}
-
-console.log('\n55. Provider — amazon (amazon.jobs search.json)');
-
-try {
-  const amazon = (await import(pathToFileURL(join(ROOT, 'providers/amazon.mjs')).href)).default;
-
-  if (amazon.id === 'amazon') pass('amazon.id is "amazon"');
-  else fail(`amazon.id is ${JSON.stringify(amazon.id)}`);
-
-  // detect() — host match, not path-spoof, https + non-string safe
-  if (amazon.detect({ name: 'X', careers_url: 'https://www.amazon.jobs/en/search' })) pass('amazon.detect() claims an amazon.jobs URL');
-  else fail('amazon.detect() should claim amazon.jobs');
-  if (amazon.detect({ name: 'X', careers_url: 'https://evil.example/www.amazon.jobs/x' }) === null) pass('amazon.detect() rejects a path-spoofed URL');
-  else fail('amazon.detect() must reject path-spoofed URLs');
-  if (amazon.detect({ name: 'X', careers_url: 42 }) === null) pass('amazon.detect() returns null for non-string careers_url');
-  else fail('amazon.detect() should return null for non-string careers_url');
-
-  // fetch() with a mock ctx — captures the request URL (to assert facet
-  // bracket-encoding) and returns a canned page, exercising the real mapping.
-  const calls = [];
-  const page1 = {
-    jobs: [
-      { title: '  Automation Engineer  ', job_path: '/en/jobs/111/automation-engineer', normalized_location: 'Erfurt, Thuringia, DEU', posted_date: 'July  1, 2026', updated_time: '10 minutes', company_name: 'Amazon' },
-      { title: 'SDE', job_path: 'https://www.amazon.jobs/en/jobs/222/sde', location: 'Berlin, DEU', posted_date: 'June 29, 2026' },
-      { title: 'No Path', normalized_location: 'X' }, // dropped — no job_path
-    ],
-  };
-  const mockCtx = {
-    transport: 'http',
-    async fetchJson(url) { calls.push(url); return calls.length === 1 ? page1 : { jobs: [] }; },
-    async fetchText() { return ''; },
-  };
-  const jobs = await amazon.fetch({ name: 'Amazon', amazon: { normalized_country_code: ['DEU'], base_query: 'engineer' } }, mockCtx);
-
-  if (jobs.length === 2) pass('amazon.fetch maps valid jobs, drops job_path-less entries');
-  else fail(`amazon.fetch returned ${jobs.length} jobs, expected 2`);
-  if (calls[0] && calls[0].includes('normalized_country_code%5B%5D=DEU')) pass('amazon.fetch bracket-encodes array facets (normalized_country_code[]=DEU)');
-  else fail(`amazon.fetch facet encoding wrong: ${calls[0]}`);
-  if (calls[0] && calls[0].includes('result_limit=100')) pass('amazon.fetch requests result_limit=100');
-  else fail('amazon.fetch should set result_limit=100');
-  const j1 = jobs.find((j) => j.url.includes('/111/'));
-  if (j1 && j1.title === 'Automation Engineer') pass('amazon.fetch trims the title');
-  else fail(`amazon.fetch title wrong: ${JSON.stringify(j1 && j1.title)}`);
-  if (j1 && j1.url === 'https://www.amazon.jobs/en/jobs/111/automation-engineer') pass('amazon.fetch builds an absolute URL from job_path');
-  else fail(`amazon.fetch url wrong: ${JSON.stringify(j1 && j1.url)}`);
-  if (j1 && j1.postedAt === Date.parse('July 1, 2026')) pass('amazon.fetch parses posted_date (ignores relative updated_time)');
-  else fail(`amazon.fetch postedAt wrong: ${JSON.stringify(j1 && j1.postedAt)}`);
-  const j2 = jobs.find((j) => j.url.includes('/222/'));
-  if (j2 && j2.url === 'https://www.amazon.jobs/en/jobs/222/sde') pass('amazon.fetch keeps an already-absolute job_path');
-  else fail(`amazon.fetch absolute url wrong: ${JSON.stringify(j2 && j2.url)}`);
-} catch (e) {
-  fail(`amazon provider tests crashed: ${e.message}`);
-}
-
-console.log('\n56. Provider — avature (career-site SearchJobs parser)');
-
-try {
-  const avature = (await import(pathToFileURL(join(ROOT, 'providers/avature.mjs')).href)).default;
-  const { parseArticles } = await import(pathToFileURL(join(ROOT, 'providers/avature.mjs')).href);
-
-  if (avature.id === 'avature') pass('avature.id is "avature"');
-  else fail(`avature.id is ${JSON.stringify(avature.id)}`);
-
-  if (avature.detect({ name: 'X', careers_url: 'https://acme.avature.net/careers/SearchJobs' })) pass('avature.detect() claims a *.avature.net URL');
-  else fail('avature.detect() should claim *.avature.net');
-  if (avature.detect({ name: 'X', careers_url: 'https://evil.example/x.avature.net/y' }) === null) pass('avature.detect() rejects a path-spoofed URL');
-  else fail('avature.detect() must reject path-spoofed URLs');
-
-  // parseArticles — a compact fragment: one article with a locale-prefixed
-  // JobDetail path + a posted date, one with no JobDetail link (dropped).
-  const origin = 'https://acme.avature.net';
-  const fragment = `
-    <article class="article article--result" id="article--1">
-      <div class="article__header"><div class="article__header__text">
-        <h3 class="title"><a class="link" href="https://acme.avature.net/careers/JobDetail/Senior-PLM-Engineer-17304/17304?businessTitle=PLM">Senior PLM Engineer &amp; Architect</a></h3>
-        <div class="article__header__text__subtitle"><span class="list-item-jobId">Job ID 17304</span><span class="list-item-posted">Posted 02-May-2026</span></div>
-      </div></div>
-    </article>
-    <article class="article article--result" id="article--2">
-      <h3 class="title"><a class="link" href="/en_US/searchjobs/JobDetail/Data-Engineer-900/900">Data Engineer</a></h3>
-      <span class="list-item-location">Munich, Germany</span>
-    </article>
-    <article class="article article--result" id="article--3">
-      <h3 class="title"><span>No link here</span></h3>
-    </article>`;
-  const arts = parseArticles(fragment, origin);
-
-  if (arts.length === 2) pass('parseArticles returns 2 articles (link-less one dropped)');
-  else fail(`parseArticles returned ${arts.length}, expected 2`);
-  const a1 = arts.find((a) => a.id === '17304');
-  if (a1 && a1.title === 'Senior PLM Engineer & Architect') pass('parseArticles decodes the title entity');
-  else fail(`parseArticles title wrong: ${JSON.stringify(a1 && a1.title)}`);
-  if (a1 && a1.url === 'https://acme.avature.net/careers/JobDetail/Senior-PLM-Engineer-17304/17304?businessTitle=PLM') pass('parseArticles keeps the absolute JobDetail URL');
-  else fail(`parseArticles url wrong: ${JSON.stringify(a1 && a1.url)}`);
-  if (a1 && a1.postedAt === Date.UTC(2026, 4, 2)) pass('parseArticles parses "Posted 02-May-2026"');
-  else fail(`parseArticles postedAt wrong: ${JSON.stringify(a1 && a1.postedAt)}`);
-  const a2 = arts.find((a) => a.id === '900');
-  if (a2 && a2.url === 'https://acme.avature.net/en_US/searchjobs/JobDetail/Data-Engineer-900/900') pass('parseArticles resolves a relative locale-prefixed JobDetail path');
-  else fail(`parseArticles relative url wrong: ${JSON.stringify(a2 && a2.url)}`);
-  if (a2 && a2.location === 'Munich, Germany') pass('parseArticles extracts a rendered location when present');
-  else fail(`parseArticles location wrong: ${JSON.stringify(a2 && a2.location)}`);
-  if (parseArticles('<div>no articles</div>', origin).length === 0) pass('parseArticles returns [] when no articles present');
-  else fail('parseArticles should return [] for markup with no articles');
-
-  // Tenant markup variants: Siemens appends a position index to the result
-  // class ("article--result 1"); Rohde & Schwarz renders the title anchor with
-  // no class="link". Both must still parse. (Regressions found on live tenants.)
-  const variants = `
-    <article class="article article--result 1" id="article--1">
-      <h3 class="title"><a class="link" href="https://acme.avature.net/en_US/externaljobs/JobDetail/Head-of-PLM/511918">Head of PLM</a></h3>
-    </article>
-    <article class="article article--result" id="article--2">
-      <h3 class="title"><a href="https://acme.avature.net/en_US/careers/JobDetail/Director-Platform/13672">Director Platform Engineering</a></h3>
-    </article>`;
-  const vArts = parseArticles(variants, origin);
-  const vSuffix = vArts.find((a) => a.id === '511918');
-  if (vSuffix && vSuffix.title === 'Head of PLM') pass('parseArticles handles the "article--result 1" class suffix (Siemens)');
-  else fail(`parseArticles missed the class-suffix variant: ${JSON.stringify(vArts.map((a) => a.id))}`);
-  const vNoClass = vArts.find((a) => a.id === '13672');
-  if (vNoClass && vNoClass.title === 'Director Platform Engineering') pass('parseArticles falls back to a JobDetail anchor without class="link" (Rohde & Schwarz)');
-  else fail(`parseArticles missed the no-class-link variant: ${JSON.stringify(vArts.map((a) => a.id))}`);
-
-  // Pagination key — default `jobOffset`, self-heals to `offset` for tenants
-  // that ignore it (Siemens). Mock fetchText with an article-less page so
-  // fetch() stops after one request and we can read the URL it built.
-  const captureFirstUrl = async (entry) => {
-    let firstUrl;
-    const ctx = { sleep: async () => {}, fetchText: async (url) => { if (firstUrl === undefined) firstUrl = url; return '<div>no articles</div>'; } };
-    await avature.fetch(entry, ctx);
-    return firstUrl;
-  };
-  const base = 'https://acme.avature.net/careers/SearchJobs';
-  if (await captureFirstUrl({ name: 'X', api: base }) === `${base}?jobOffset=0`) pass('avature.fetch() defaults pagination to ?jobOffset=N');
-  else fail('avature.fetch() should default to jobOffset');
-  if (await captureFirstUrl({ name: 'X', api: base, offset_param: 'offset' }) === `${base}?offset=0`) pass('avature.fetch() honours offset_param override (Siemens: ?offset=N)');
-  else fail('avature.fetch() should use offset_param when set');
-  if (await captureFirstUrl({ name: 'X', api: base, offset_param: '  ' }) === `${base}?jobOffset=0`) pass('avature.fetch() falls back to jobOffset for a blank offset_param');
-  else fail('avature.fetch() should ignore a blank offset_param');
-
-  // Self-heal — jobOffset→offset when the primary key is inert.
-  const originB = 'https://acme.avature.net';
-  const mkHtml = (ids) => ids.map((id) =>
-    `<article class="article article--result"><h3 class="title"><a class="link" href="${originB}/careers/JobDetail/Role-${id}/${id}">Role ${id}</a></h3></article>`).join('');
-  // Build a ctx whose fetchText answers from a {param: (pageIndex)=>ids} map.
-  const mkCtx = () => {
-    const calls = [];
-    let sleeps = 0;
-    const ctx = {
-      sleep: async () => { sleeps += 1; },
-      fetchText: async (url) => {
-        calls.push(url);
-        const u = new URL(url);
-        const jo = u.searchParams.get('jobOffset');
-        const of = u.searchParams.get('offset');
-        if (jo !== null) return mkHtml([1, 2, 3, 4, 5, 6]); // jobOffset inert: always page 0
-        if (of !== null) {
-          const n = Number(of) / 6;
-          if (n === 0) return mkHtml([1, 2, 3, 4, 5, 6]);
-          if (n === 1) return mkHtml([7, 8, 9, 10, 11, 12]); // offset advances
-          if (n === 2) return mkHtml([13, 14]); // partial → last page
-          return mkHtml([]);
-        }
-        return '<div>no articles</div>';
-      },
-    };
-    return { ctx, calls, sleeps: () => sleeps };
-  };
-  const heal = mkCtx();
-  const healed = await avature.fetch({ name: 'X', api: base }, heal.ctx);
-  if (healed.length === 14) pass('avature.fetch() self-heals jobOffset→offset and walks the full board (14 jobs)');
-  else fail(`avature.fetch() self-heal wrong count: ${healed.length} (expected 14)`);
-  if (heal.calls.some((u) => /[?&]offset=/.test(u))) pass('avature.fetch() self-heal retries with ?offset=N');
-  else fail('avature.fetch() self-heal should retry with offset');
-  if (heal.sleeps() > 0) pass('avature.fetch() throttles between pages (sleep called)');
-  else fail('avature.fetch() should sleep between pages');
-
-  // No self-heal when jobOffset works: offset= must never be requested.
-  const workingCtx = {
-    calls: [],
-    sleep: async () => {},
-    fetchText: async function (url) {
-      this.calls.push(url);
-      const n = Number(new URL(url).searchParams.get('jobOffset')) / 6;
-      if (n === 0) return mkHtml([1, 2, 3, 4, 5, 6]);
-      if (n === 1) return mkHtml([7, 8, 9, 10, 11, 12]);
-      return mkHtml([13, 14]); // last (partial)
-    },
-  };
-  const worked = await avature.fetch({ name: 'X', api: base }, workingCtx);
-  if (worked.length === 14 && workingCtx.calls.every((u) => /[?&]jobOffset=/.test(u))) pass('avature.fetch() does not self-heal when jobOffset already advances');
-  else fail(`avature.fetch() spurious self-heal: ${worked.length} jobs, calls ${JSON.stringify(workingCtx.calls.map((u) => u.split('?')[1]))}`);
-
-  // Self-heal must fire even when the inert primary key returns an EMPTY page 1
-  // (not a repeat of page 0) — the empty-page break must not pre-empt the heal.
-  const emptyP1Ctx = {
-    calls: [],
-    sleep: async () => {},
-    fetchText: async function (url) {
-      this.calls.push(url);
-      const u = new URL(url);
-      const jo = u.searchParams.get('jobOffset');
-      const of = u.searchParams.get('offset');
-      if (jo !== null) return Number(jo) === 0 ? mkHtml([1, 2, 3, 4, 5, 6]) : mkHtml([]); // page 1+ empty
-      if (of !== null) {
-        const n = Number(of) / 6;
-        if (n === 1) return mkHtml([7, 8, 9, 10, 11, 12]);
-        if (n === 2) return mkHtml([13, 14]);
-        return mkHtml([]);
-      }
-      return '<div>no articles</div>';
-    },
-  };
-  const emptyHealed = await avature.fetch({ name: 'X', api: base }, emptyP1Ctx);
-  if (emptyHealed.length === 14 && emptyP1Ctx.calls.some((u) => /[?&]offset=/.test(u))) pass('avature.fetch() self-heals when the inert key returns an empty page 1');
-  else fail(`avature.fetch() failed to heal empty page 1: ${emptyHealed.length} jobs`);
-} catch (e) {
-  fail(`avature provider tests crashed: ${e.message}`);
 }
 
 console.log('\n57. Provider — dassault (Exalead card_search_api XML parser)');
